refactor(billing): improve code layout
