+++ app/(home)/premium/page.tsx
@@ -1,17 +1,15 @@
 import Comment from "@/components/comment";
 import { OverviewLink } from "@/components/overview-link";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Anchor } from "@/components/ui/typography";
 import BotStylePic from "@/public/docs-assets/bot-style.webp";
-import MrBeastPic from "@/public/mrbeast.webp";
 import { cn } from "@/utils/cn";
+import { MONTHLY_PRICES } from "@/utils/premium";
 import { getBaseUrl, getCanonicalUrl } from "@/utils/urls";
 import type { Metadata } from "next";
 import { Lexend } from "next/font/google";
 import Image from "next/image";
 import Link from "next/link";
-import type { ReactNode } from "react";
 import { BsDiscord } from "react-icons/bs";
 import {
     HiBell,
@@ -32,15 +30,11 @@ import {
     HiVolumeUp
 } from "react-icons/hi";
 
-import { GiftBanner } from "./gift-banner";
-import { Subscribe } from "./subscribe.component";
+import type { Feature } from "./card";
+import { PricingCard } from "./card";
+import { Subscribe } from "./subscribe";
 
 const lexend = Lexend({ subsets: ["latin"] });
-
-interface Feature {
-    label: string;
-    icon: ReactNode;
-}
 
 const freeFeatures: Feature[] = [
     { label: "Text to Speech", icon: <HiVolumeUp /> },
@@ -97,79 +91,11 @@ export const generateMetadata = (): Metadata => {
     };
 };
 
-interface PricingCardProps {
-    title: string;
-    price: string;
-    priceSuffix: string;
-    features: Feature[];
-    accentColor: string;
-    badge?: string;
-    highlighted?: boolean;
-    action: ReactNode;
-}
-
-function PricingCard({ title, price, priceSuffix, features, accentColor, badge, highlighted, action }: PricingCardProps) {
-    return (
-        <div className="relative w-full md:max-w-md">
-            {badge && (
-                <Badge
-                    className="absolute -top-3 left-1/2 -translate-x-1/2 z-10 bg-violet-500 text-white border-violet-500 uppercase tracking-wider text-[10px] font-bold px-3 py-1"
-                >
-                    {badge}
-                </Badge>
-            )}
-            <div className={cn(
-                "dark:bg-wamellow bg-wamellow-100 rounded-xl p-6 h-full flex flex-col duration-300",
-                highlighted
-                    ? "border border-violet-500/40 relative overflow-hidden"
-                    : "border border-transparent hover:border-violet-500/20"
-            )}>
-                {highlighted && (
-                    <div className="absolute inset-0 bg-violet-500/3 pointer-events-none" />
-                )}
-
-                <div className={cn("mb-6", highlighted && "relative")}>
-                    <h2 className={cn("text-2xl font-semibold dark:text-neutral-100 text-neutral-900 mb-1", lexend.className)}>
-                        {title}
-                    </h2>
-                    {price !== "0" && <GiftBanner />}
-                    <div className="flex items-end gap-1 mt-2">
-                        <span className="text-muted-foreground text-sm mb-1">€</span>
-                        <span className={cn("text-4xl font-bold dark:text-neutral-100 text-neutral-900", lexend.className)}>{price}</span>
-                        <span className="text-muted-foreground text-sm mb-1">/ {priceSuffix}</span>
-                    </div>
-                </div>
-
-                <div className={cn("space-y-3 flex-1", highlighted && "relative")}>
-                    {features.map((feature) => (
-                        <div key={feature.label} className="flex items-center gap-3">
-                            <span className={cn("size-5 shrink-0 flex items-center justify-center [&>svg]:size-[18px]", accentColor)}>
-                                {feature.icon}
-                            </span>
-                            <span className="dark:text-neutral-300 text-neutral-700 text-sm">{feature.label}</span>
-                        </div>
-                    ))}
-                    {price === "0" &&
-                        <Image
-                            src={MrBeastPic}
-                            alt=""
-                            className="h-65 w-auto rounded-lg mt-6 hidden md:block"
-                        />}
-                </div>
-
-                <div className={cn("mt-8", highlighted && "relative")}>
-                    {action}
-                </div>
-            </div>
-        </div>
-    );
-}
-
 export default function Home() {
     return (
         <div className="w-full">
             <h1
-                className={cn(lexend.className, "lg:text-5xl text-4xl font-bold dark:text-neutral-100 text-neutral-900 wrap-break-words mb-2 flex gap-4")}
+                className={cn(lexend.className, "md:text-5xl text-4xl font-bold dark:text-neutral-100 text-neutral-900 wrap-break-words mb-2 flex gap-4")}
             >
                 <span className="hidden md:block">Wamellow</span>
                 <span className="bg-linear-to-r from-indigo-400 to-pink-400 bg-clip-text text-transparent break-keep">Premium</span>
@@ -207,7 +133,7 @@ export default function Home() {
 
                 <PricingCard
                     title="Wamellow Premium"
-                    price="4"
+                    price={MONTHLY_PRICES[0].toString()}
                     priceSuffix="per month"
                     features={premiumFeatures}
                     accentColor="text-violet-400"
